requirements.txt
Django==4.2.8
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
twilio==8.10.0
django-crispy-forms==2.1
crispy-bootstrap5==2.0.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.26.3
python-dateutil==2.8.2
requests==2.31.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
